+++ derive_builder/src/options/darling_opts.rs
@@ -43,6 +43,8 @@ pub struct BuildFn {
     skip: bool,
     name: Ident,
     validate: Option<Path>,
+    public: Flag,
+    private: Flag,
 }
 
 impl Default for BuildFn {
@@ -51,7 +53,19 @@ impl Default for BuildFn {
             skip: false,
             name: Ident::from("build"),
             validate: None,
+            public: Default::default(),
+            private: Default::default(),
         }
+    }
+}
+
+impl FlagVisibility for BuildFn {
+    fn public(&self) -> &Flag {
+        &self.public
+        }
+
+    fn private(&self) -> &Flag {
+        &self.private
     }
 }
 
@@ -282,8 +296,19 @@ impl Options {
             .expect("Struct name with Builder suffix should be an ident")
     }
 
+    /// The visibility of the builder struct.
+    /// If a visibility was declared in attributes, that will be used;
+    /// otherwise the struct's own visibility will be used.
     pub fn builder_vis(&self) -> Visibility {
-        self.vis.clone()
+        self.as_expressed_vis().unwrap_or_else(|| self.vis.clone())
+    }
+
+    /// Get the visibility of the emitted `build` method.
+    /// This defaults to the visibility of the parent builder, but can be overridden.
+    pub fn build_method_vis(&self) -> Visibility {
+        self.build_fn
+            .as_expressed_vis()
+            .unwrap_or_else(|| self.builder_vis())
     }
 
     pub fn raw_fields<'a>(&'a self) -> Vec<&'a Field> {
@@ -345,7 +370,7 @@ impl Options {
         BuildMethod {
             enabled: !self.build_fn.skip,
             ident: &self.build_fn.name,
-            visibility: self.builder_vis(),
+            visibility: self.build_method_vis(),
             pattern: self.pattern,
             target_ty: &self.ident,
             target_ty_generics: Some(ty_generics),